PTB BETA RELEASE "Broken Windows"

This is PTB BETA "Broken Windows"

Highlights:

64-Bit Matlab support for Microsoft Windows added.
32-Bit Octave support for OSX removed, 64-Bit Octave 3.6 support for OSX added.
Octave support for Windows removed.
Quicktime support for Windows removed.
GStreamer support for 32-Bit Matlab on OSX added.

Massive movie playback performance and functionality improvements for GStreamer based movie playback engine. Some perf. improvements for videocapture engine, and basic GeniCam/GigE video capture support.

Support for GL_TEXTURE_2D textures as rendertargets (e.g., Offscreen windows and 'TransformTexture' etc.) -- For improved OpenGL interop with 3rd party OpenGL code. Non-power-of-two texture support. Mipmap filtering support and demos for efficient blurring via gpu accelerated resolution pyramids.

Linux: Windowed window support and GUI window support brought to same level of functionality and behaviour as on OSX.

Various improvements and bug fixes.

PsychtoolboxKernelDriver on OSX now supports Intel IGP's as well.
Hybrid graphics dual-gpu configs now supported on OSX for our low-level code.

+++ PsychSourceGL/Source/Common/Screen/SCREENFramePoly.c
@@ -44,9 +44,9 @@ PsychError SCREENFramePoly(void)
 	
 	PsychColorType					color;
 	PsychWindowRecordType			*windowRecord;
-	int								whiteValue;
+	double							whiteValue;
 	int								i, mSize, nSize, pSize;
-	psych_bool							isArgThere;
+	psych_bool						isArgThere;
 	double							penSize, *pointList;
     
 	//all sub functions should have these two lines
